requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.6
django-bootstrap4==2.3.1
django-jquery==3.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
emoji==1.2.0
et-xmlfile==1.0.1
fasttext==0.9.2
hazm==0.7.0
jdcal==1.4.1
libwapiti==0.2.1
nltk==3.3
numpy==1.20.0
openpyxl==3.0.6
pandas==1.2.1
psycopg2-binary==2.8.6
pybind11==2.6.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
xlrd==2.0.1
XlsxWriter==1.3.7
